ingestion/requirements.txt
langchain-core
langchain-community
langchain-openai
chromadb
transformers
torch
watchdog
markdown
